// src/api/generated/schemas/templateResponse.ts
/**
 * Generated by orval v7.12.2 🍺
 * Do not edit manually.
 * Notifiable API
 * A simple example ASP.NET Core Web API
 * OpenAPI spec version: v1
 */
import type { TemplateStatus } from './templateStatus';
import type { NotificationChannelType } from './notificationChannelType';
import type { TemplateResponseContent } from './templateResponseContent';

export interface TemplateResponse {
  id?: string;
  organizationId?: string;
  code?: string;
  status?: TemplateStatus;
  channelType?: NotificationChannelType;
  language?: string;
  isExternal?: boolean;
  /** @nullable */
  externalProvider?: string | null;
  /** @nullable */
  externalTemplateId?: string | null;
  /** @nullable */
  content?: TemplateResponseContent;
  createdByUserId?: string;
  createdAt?: string;
  /** @nullable */
  modifiedByUserId?: string | null;
  /** @nullable */
  modifiedAt?: string | null;
}